// front/src/types/api.d.ts
export interface components {
    schemas: {
// ...
        Consumer: {
            readonly id?: number;
            name?: string;
            apiKey?: string;
            contactEmail?: string;
            allowedOrigins?: (string | null)[];
            readonly roles?: (string | null)[];
            readonly userIdentifier?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}